+++ src/frontend/screens/central/public-tab/asyncstorage.ts
@@ -1,15 +1,30 @@
-// SPDX-FileCopyrightText: 2018-2020 The Manyverse Authors
+// SPDX-FileCopyrightText: 2018-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
 import xs, {Stream} from 'xstream';
-import {Command} from 'cycle-native-asyncstorage';
+import {AsyncStorageSource, Command} from 'cycle-native-asyncstorage';
 
-type Actions = {
+interface Actions {
   updateSessionTimestamp$: Stream<any>;
-};
+}
 
-export default function asyncStorage(actions: Actions) {
+export default function asyncStorage(
+  actions: Actions,
+  asyncStorageSource: AsyncStorageSource,
+) {
+  const firstVisit$ = asyncStorageSource
+    .getItem('firstVisit')
+    .filter((resultStr: string | null) => !resultStr)
+    .map(
+      () =>
+        ({
+          type: 'setItem',
+          key: 'firstVisit',
+          value: `${Date.now()}`,
+        } as Command),
+    );
+
   const latestVisit$ = xs.of({
     type: 'setItem',
     key: 'latestVisit',
@@ -25,5 +40,5 @@ export default function asyncStorage(actions: Actions) {
       } as Command),
   );
 
-  return xs.merge(latestVisit$, lastSession$);
+  return xs.merge(latestVisit$, lastSession$, firstVisit$);
 }