+++ src/frontend/screens/desktop-frame/index.web.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2021-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -6,19 +6,21 @@ import xs, {Stream} from 'xstream';
 import {Command, NavSource, StackElement} from 'cycle-native-navigation';
 import {ReactElement} from 'react';
 import {ReactSource} from '@cycle/react';
+import {HTTPSource, RequestInput as HTTPReq} from '@cycle/http';
 import {Reducer, StateSource} from '@cycle/state';
 import {GlobalEvent} from '../../drivers/eventbus';
 import {SSBSource} from '../../drivers/ssb';
 import {DialogSource} from '../../drivers/dialogs';
-import MAIL_TO_BUG_REPORT from '../../components/mail-to-bug-report';
 import model, {State} from './model';
 import view from './view';
 import intent from './intent';
 import navigation from './navigation';
+import linking from '../drawer/linking';
 
 export interface Sources {
   screen: ReactSource;
   navigation: NavSource;
+  http: HTTPSource;
   navigationStack: StateSource<Array<StackElement>>;
   children: Stream<Array<ReactElement>>;
   globalEventBus: Stream<GlobalEvent>;
@@ -30,6 +32,7 @@ export interface Sources {
 export interface Sinks {
   screen: Stream<ReactElement<any>>;
   navigation: Stream<Command>;
+  http: Stream<HTTPReq>;
   state: Stream<Reducer<State>>;
   globalEventBus: Stream<GlobalEvent>;
   linking: Stream<string>;
@@ -38,7 +41,7 @@ export interface Sinks {
 export function desktopFrame(sources: Sources): Sinks {
   const state$ = sources.state.stream;
 
-  const actions = intent(sources.screen, sources.dialog, state$);
+  const actions = intent(sources.screen, sources.dialog, sources.http, state$);
 
   const event$ = xs.merge(
     actions.changeTab$.map(
@@ -72,15 +75,23 @@ export function desktopFrame(sources: Sources): Sinks {
 
   const command$ = navigation(actions, state$, sources.navigationStack.stream);
 
-  const linking$ = xs.merge(
-    actions.emailBugReport$.mapTo(MAIL_TO_BUG_REPORT),
-    actions.openTranslate$.mapTo('https://www.manyver.se/translations/'),
+  const linking$ = linking(actions);
+
+  const httpReq$ = actions.checkNewVersion$.map(
+    () =>
+      ({
+        url: 'https://manyver.se/latestversion',
+        method: 'GET',
+        accept: 'application/json',
+        category: 'latestversion',
+      } as HTTPReq),
   );
 
   return {
     screen: vdom$,
     state: reducer$,
     navigation: command$,
+    http: httpReq$,
     globalEventBus: event$,
     linking: linking$,
   };