+++ src/frontend/screens/desktop-frame/model.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2021-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -18,6 +18,7 @@ import progressCalculation, {
   State as ProgressState,
   INITIAL_STATE as INITIAL_PROGRESS_STATE,
 } from '../../components/progressCalculation';
+import currentVersion from '../../versionName';
 
 export interface State extends ProgressState {
   selfFeedId: FeedId;
@@ -34,10 +35,13 @@ export interface State extends ProgressState {
   numOfPublicUpdates: number;
   numOfPrivateUpdates: number;
   numOfActivityUpdates: number;
+  allowCheckingNewVersion: boolean;
+  hasNewVersion: boolean;
 }
 
 interface Actions {
   changeTab$: Stream<State['currentTab']>;
+  latestVersionResponse$: Stream<string>;
 }
 
 export default function model(
@@ -75,6 +79,8 @@ export default function model(
             numOfPublicUpdates: 0,
             numOfPrivateUpdates: 0,
             numOfActivityUpdates: 0,
+            allowCheckingNewVersion: false,
+            hasNewVersion: false,
             ...INITIAL_PROGRESS_STATE,
           };
         } else {
@@ -101,6 +107,8 @@ export default function model(
               numOfPublicUpdates: 0,
               numOfPrivateUpdates: 0,
               numOfActivityUpdates: 0,
+              allowCheckingNewVersion: false,
+              hasNewVersion: false,
               ...INITIAL_PROGRESS_STATE,
             };
           } else {
@@ -173,6 +181,43 @@ export default function model(
       },
   );
 
+  const readSettingsReducer$ = ssbSource.readSettings().map(
+    (settings) =>
+      function readSettingsReducer(prev: State): State {
+        return {
+          ...prev,
+          allowCheckingNewVersion: settings.allowCheckingNewVersion ?? false,
+        };
+      },
+  );
+
+  const allowCheckingNewVersionReducer$ = globalEventBus
+    .filter((ev) => ev.type === 'approveCheckingNewVersion')
+    .map(
+      () =>
+        function allowCheckingNewVersionReducer(prev: State): State {
+          return {...prev, allowCheckingNewVersion: true};
+        },
+    );
+
+  const hasNewVersionReducer$ = actions.latestVersionResponse$
+    .map((latestVersion) => {
+      const [majorPrev, minorPrev, restPrev] = currentVersion.split('.');
+      const [majorNext, minorNext, restNext] = latestVersion.split('.');
+      const patchPrev = restPrev.split('-')[0];
+      const patchNext = restNext.split('-')[0];
+      if (majorNext > majorPrev) return true;
+      if (minorNext > minorPrev) return true;
+      if (patchNext > patchPrev) return true;
+      return false;
+    })
+    .map(
+      (hasNewVersion) =>
+        function hasNewVersionReducer(prev: State): State {
+          return {...prev, hasNewVersion};
+        },
+    );
+
   const progressReducer$ = progressCalculation(ssbSource) as Stream<
     Reducer<State>
   >;
@@ -185,6 +230,9 @@ export default function model(
     updatePrivateCounterReducer$,
     updateActivityCounterReducer$,
     updateConnectionsReducer$,
+    readSettingsReducer$,
+    allowCheckingNewVersionReducer$,
+    hasNewVersionReducer$,
     progressReducer$,
   );
 }