+++ src/frontend/screens/drawer/linking.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020-2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2020-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -8,12 +8,15 @@ import MAIL_TO_BUG_REPORT from '../../components/mail-to-bug-report';
 interface Actions {
   emailBugReport$: Stream<any>;
   openTranslate$: Stream<any>;
+  downloadNewVersion$: Stream<any>;
 }
 
 export default function linking(actions: Actions) {
   return xs.merge(
     actions.emailBugReport$.mapTo(MAIL_TO_BUG_REPORT),
 
-    actions.openTranslate$.mapTo('https://www.manyver.se/translations/'),
+    actions.openTranslate$.mapTo('https://manyver.se/translations'),
+
+    actions.downloadNewVersion$.mapTo('https://manyver.se/download'),
   );
 }