+++ src/frontend/screens/drawer/model.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2018-2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2018-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -6,19 +6,31 @@ import xs, {Stream} from 'xstream';
 import {FeedId} from 'ssb-typescript';
 import {Reducer} from '@cycle/state';
 import {SSBSource} from '../../drivers/ssb';
+import {GlobalEvent} from '../../drivers/eventbus';
 import progressCalculation, {
   State as ProgressState,
   INITIAL_STATE as INITIAL_PROGRESS_STATE,
 } from '../../components/progressCalculation';
+import currentVersion from '../../versionName';
 
 export interface State extends ProgressState {
   selfFeedId: FeedId;
   canPublishSSB: boolean;
+  allowCheckingNewVersion: boolean;
+  hasNewVersion: boolean;
   selfAvatarUrl?: string;
   name?: string;
 }
 
-export default function model(ssbSource: SSBSource): Stream<Reducer<State>> {
+interface Actions {
+  latestVersionResponse$: Stream<string>;
+}
+
+export default function model(
+  actions: Actions,
+  ssbSource: SSBSource,
+  globalEventBus: Stream<GlobalEvent>,
+): Stream<Reducer<State>> {
   const selfFeedId$ = ssbSource.selfFeedId$.take(1);
 
   const selfFeedIdReducer$ = selfFeedId$.map(
@@ -28,6 +40,8 @@ export default function model(ssbSource: SSBSource): Stream<Reducer<State>> {
           return {
             selfFeedId,
             canPublishSSB: true,
+            allowCheckingNewVersion: false,
+            hasNewVersion: false,
             ...INITIAL_PROGRESS_STATE,
           };
         } else {
@@ -52,6 +66,8 @@ export default function model(ssbSource: SSBSource): Stream<Reducer<State>> {
               selfAvatarUrl: about.imageUrl,
               name,
               canPublishSSB: true,
+              allowCheckingNewVersion: false,
+              hasNewVersion: false,
               ...INITIAL_PROGRESS_STATE,
             };
           } else {
@@ -64,9 +80,53 @@ export default function model(ssbSource: SSBSource): Stream<Reducer<State>> {
         },
     );
 
+  const readSettingsReducer$ = ssbSource.readSettings().map(
+    (settings) =>
+      function readSettingsReducer(prev: State): State {
+        return {
+          ...prev,
+          allowCheckingNewVersion: settings.allowCheckingNewVersion ?? false,
+        };
+      },
+  );
+
+  const allowCheckingNewVersionReducer$ = globalEventBus
+    .filter((ev) => ev.type === 'approveCheckingNewVersion')
+    .map(
+      () =>
+        function allowCheckingNewVersionReducer(prev: State): State {
+          return {...prev, allowCheckingNewVersion: true};
+        },
+    );
+
+  const hasNewVersionReducer$ = actions.latestVersionResponse$
+    .map((latestVersion) => {
+      const [majorPrev, minorPrev, restPrev] = currentVersion.split('.');
+      const [majorNext, minorNext, restNext] = latestVersion.split('.');
+      const patchPrev = restPrev.split('-')[0];
+      const patchNext = restNext.split('-')[0];
+      if (majorNext > majorPrev) return true;
+      if (minorNext > minorPrev) return true;
+      if (patchNext > patchPrev) return true;
+      return false;
+    })
+    .map(
+      (hasNewVersion) =>
+        function hasNewVersionReducer(prev: State): State {
+          return {...prev, hasNewVersion};
+        },
+    );
+
   const progressReducer$ = progressCalculation(ssbSource) as Stream<
     Reducer<State>
   >;
 
-  return xs.merge(selfFeedIdReducer$, aboutReducer$, progressReducer$);
+  return xs.merge(
+    selfFeedIdReducer$,
+    aboutReducer$,
+    progressReducer$,
+    readSettingsReducer$,
+    allowCheckingNewVersionReducer$,
+    hasNewVersionReducer$,
+  );
 }