+++ src/frontend/screens/global/intent.ts
@@ -1,8 +1,10 @@
-// SPDX-FileCopyrightText: 2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2021-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
 import xs, {Stream} from 'xstream';
+import delay from 'xstream/extra/delay';
+import {NativeModules, Platform} from 'react-native';
 import {
   isFeedSSBURI,
   isMessageSSBURI,
@@ -19,9 +21,10 @@ import {
   TriggerMsgCypherlink,
 } from '../../drivers/eventbus';
 import {SSBSource} from '../../drivers/ssb';
-import {DialogSource} from '../../drivers/dialogs';
+import {AlertAction, DialogSource} from '../../drivers/dialogs';
 import {t} from '../../drivers/localization';
 import {Palette} from '../../global-styles/palette';
+import {State} from './model';
 const Ref = require('ssb-ref');
 const urlParse = require('url-parse');
 
@@ -30,8 +33,50 @@ export default function intent(
   linkingSource: Stream<string>,
   dialogSource: DialogSource,
   ssbSource: SSBSource,
+  state$: Stream<State>,
 ) {
   const canNowHandleLinks$ = ssbSource.connStarted$;
+
+  let responseCheckingNewVersion$: Stream<AlertAction>;
+  if (
+    Platform.OS === 'web' ||
+    (Platform.OS === 'android' && NativeModules.BuildConfig.FLAVOR === 'indie')
+  ) {
+    const DURATION_CHECK_NEW_VERSION = 1000 * 60 * 60 * 24; // 1 day
+    responseCheckingNewVersion$ = state$
+      .map(
+        (state) =>
+          state.allowCheckingNewVersion === null &&
+          !!state.firstVisit &&
+          state.firstVisit + DURATION_CHECK_NEW_VERSION < Date.now(),
+      )
+      .filter((showDialog) => showDialog === true)
+      .take(1)
+      .compose(delay(5e3))
+      .map(() =>
+        dialogSource.alert(
+          t('drawer.dialogs.update.title'),
+          t('drawer.dialogs.update.description'),
+          {
+            ...Palette.dialogColors,
+            positiveText: t('call_to_action.yes'),
+            negativeText: t('call_to_action.no'),
+            markdownOnDesktop: true,
+          },
+        ),
+      )
+      .flatten();
+  } else {
+    responseCheckingNewVersion$ = xs.never();
+  }
+
+  const approvedCheckingNewVersion$ = responseCheckingNewVersion$.filter(
+    (res) => res.action === 'actionPositive',
+  );
+
+  const rejectedCheckingNewVersion$ = responseCheckingNewVersion$.filter(
+    (res) => res.action === 'actionNegative',
+  );
 
   const link$ = canNowHandleLinks$
     .map(() => linkingSource)
@@ -126,6 +171,8 @@ export default function intent(
     handleUriClaimInvite$,
     handleUriConsumeAlias$,
     handleUriStartHttpAuth$,
+    approvedCheckingNewVersion$,
+    rejectedCheckingNewVersion$,
     connectToPeer$,
     confirmedSignInRoom$,
     goToProfile$,