+++ src/frontend/screens/global/model.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2020-2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2020-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -7,11 +7,20 @@ import {FeedId} from 'ssb-typescript';
 import {AsyncStorageSource} from 'cycle-native-asyncstorage';
 import {SSBSource} from '../../drivers/ssb';
 
-export type State = {
+export interface State {
+  firstVisit?: number;
   lastSessionTimestamp?: number;
   selfFeedId?: FeedId;
   selfAvatarUrl?: string;
-};
+
+  /**
+   * - `undefined` means "not yet loaded"
+   * - `null` means loaded but the user has not yet made a choice
+   * - `true` means the user chose to allow checking new versions
+   * - `false` means the user chose to reject checking new versions
+   */
+  allowCheckingNewVersion: boolean | null | undefined;
+}
 
 export default function model(
   ssbSource: SSBSource,
@@ -23,25 +32,72 @@ export default function model(
       ssbSource.profileAbout$(selfFeedId).map(
         (about) =>
           function aboutReducer(prev?: State): State {
-            return {...(prev ?? {}), selfFeedId, selfAvatarUrl: about.imageUrl};
+            const initial = prev ?? {allowCheckingNewVersion: undefined};
+            return {
+              ...initial,
+              selfFeedId,
+              selfAvatarUrl: about.imageUrl,
+            };
           },
       ),
     )
     .flatten();
+
+  const firstVisitReducer$ = asyncStorageSource
+    .getItem('firstVisit')
+    .filter((resultStr: string | null) => !!resultStr)
+    .map(
+      (resultStr: string) =>
+        function firstVisitReducer(prev?: State): State {
+          const initial = prev ?? {allowCheckingNewVersion: undefined};
+          const firstVisit = parseInt(resultStr, 10);
+          if (isNaN(firstVisit)) {
+            return initial;
+          } else {
+            return {...initial, firstVisit};
+          }
+        },
+    );
 
   const lastSessionTimestampReducer$ = asyncStorageSource
     .getItem('lastSessionTimestamp')
     .map(
       (resultStr) =>
         function lastSessionTimestampReducer(prev?: State): State {
+          const initial = prev ?? {allowCheckingNewVersion: undefined};
           const lastSessionTimestamp = parseInt(resultStr ?? '', 10);
           if (isNaN(lastSessionTimestamp)) {
-            return prev ?? {};
+            return initial;
           } else {
-            return {...prev, lastSessionTimestamp};
+            return {
+              ...initial,
+              lastSessionTimestamp,
+            };
           }
         },
     );
 
-  return xs.merge(aboutReducer$, lastSessionTimestampReducer$);
+  const readSettingsReducer$ = ssbSource.readSettings().map(
+    (settings) =>
+      function readSettingsReducer(prev?: State): State {
+        if (typeof settings.allowCheckingNewVersion === 'boolean') {
+          return {
+            ...prev,
+            allowCheckingNewVersion: settings.allowCheckingNewVersion,
+          };
+        } else {
+          return {
+            ...prev,
+            allowCheckingNewVersion: null,
+          };
+        }
+      },
+  );
+
+  return xs.merge(
+    aboutReducer$,
+    firstVisitReducer$,
+    lastSessionTimestampReducer$,
+    readSettingsReducer$,
+  );
 }