+++ src/frontend/ssb/manifest.ts
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2018-2021 The Manyverse Authors
+// SPDX-FileCopyrightText: 2018-2022 The Manyverse Authors
 //
 // SPDX-License-Identifier: MPL-2.0
 
@@ -193,6 +193,7 @@ export default {
     updateBlobsPurge: 'sync',
     updateShowFollows: 'sync',
     updateDetailedLogs: 'sync',
+    updateAllowCheckingNewVersion: 'sync',
   },
   syncing: {
     migrating: 'source',