+++ docs/releases.md
@@ -91,7 +91,9 @@ Then make an Electron release:
 
 - Deploy as an APK on the website
   - Go to `android/app/build/outputs/apk/indie/release` and copy the `app-indie-release.apk` file
-  - Put the file in the manyverse-website repo and deploy the website
+  - Put the file in the manyverse-website repo at `./static/app-indie-release.apk`
+  - Update the versionName at `./static/latestversion.json`
+  - Deploy the website
 - Publish Android to Google Play
   - Take the APK file from `android/app/build/outputs/apk/googlePlay/release`, upload and publish it on Google Play developer website
 - Publish iOS to the App Store